Implement a KohnShamConvergenceError
If this error is caught as a special type of RuntimeError, i.e. the KohnShamConvergenceError, it becomes possible to programmatically handle what happens once this error is caught. For example, the input script can retry the calculation using a smaller mixing parameter. This type of code is modeled exactly on GPAW (https://gitlab.com/gpaw/gpaw/blob/master/gpaw/__init__.py).